% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbor.R
\name{neighbor}
\alias{neighbor}
\title{Find nearest neighbors}
\usage{
neighbor(neighbor_size, coords, obs, ref = TRUE, nonref = NULL)
}
\arguments{
\item{neighbor_size}{neighborhood size (number of nearest-neighbors, say \eqn{L})}

\item{coords}{an \eqn{n \times 2} matrix of reference set coordinates, e.g., longitude and latitude.}

\item{obs}{a vector of observations of length \eqn{n} associated with the matrix of reference set coordinates.}

\item{ref}{if \code{TRUE}, nearest neighbors of reference set are searched. If \code{FALSE}, 
argument \code{nonref} must be specified.}

\item{nonref}{an \eqn{r \times 2} matrix of non-reference set coordinates, e.g., longitude and latitude.
This argument only needs to be specified when \code{ref = FALSE}.}
}
\value{
The function returns a list that comprises:
\tabular{ll}{

\code{ne_dist} \tab an \eqn{n \times L} matrix where the \eqn{(i,j)}-th element of the matrix
is the Euclidean distance between the \eqn{i}-th reference set location and its \eqn{j}-th nearest neighbor. \cr
\tab \cr
\code{ne_idx} \tab an \eqn{n \times L} matrix where the \eqn{(i,j)}-th element of the matrix
is the row index of the \eqn{i}-th reference set location's \eqn{j}-th nearest neighbor in \code{coords}. \cr
\tab \cr
\code{ne_obs} \tab an \eqn{n \times L} matrix where the \eqn{(i,j)}-th element of the matrix is the observation
associated with the \eqn{i}-th reference set location's \eqn{j}-th nearest neighbor in \code{coords}. \cr
}
}
\description{
This function finds nearest neighbors for locations in the reference set or in
the non-reference set. Neighbors are found in the reference set.
}
\author{
Xiaotian Zheng \email{xiaotianzheng42@gmail.com}
}
