% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnmp-generics.R
\name{pnnmp}
\alias{pnnmp}
\title{Function for the conditional cumulative distribution function at new locations}
\usage{
pnnmp(object, nonref_coords, probs)
}
\arguments{
\item{object}{an object of class nnmp.}

\item{nonref_coords}{an \eqn{m \times 2} matrix of spatial coordinates.}

\item{probs}{a vector of marginal probabilities corresponding to the NNMP response.}
}
\description{
This function produces posterior samples of the conditional cumulative distribution functions of
an NNMP for a set of new locations. The function currently supports only stationary beta/gamma NNMPs.
}
\references{
Zheng, X., Kottas, A., and Sansó, B. 
"Nearest-neighbor mixture models for non-Gaussian spatial processes". 
Bayesian Anal. 18 (4) 1191 - 1222, December 2023.
DOI: \href{https://doi.org/10.1214/23-BA1405}{https://doi.org/10.1214/23-BA1405}.
}
\author{
Xiaotian Zheng \email{xiaotianzheng42@gmail.com}
}
