% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnmp-generics.R
\name{predict.nnmp}
\alias{predict.nnmp}
\title{Function for prediction at new locations}
\usage{
\method{predict}{nnmp}(
  object,
  nonref_covars = NULL,
  nonref_coords,
  probs = c(0.025, 0.5, 0.975),
  predict_sam = FALSE,
  nonref_par_labels = NULL,
  compute_control = NULL,
  verbose = TRUE,
  nreport = NULL,
  ...
)
}
\arguments{
\item{object}{an object of class nnmp.}

\item{nonref_covars}{an \eqn{m \times p} matrix of covariates for \eqn{m} new locations.}

\item{nonref_coords}{an \eqn{m \times 2} matrix of the corresponding spatial coordinates.}

\item{probs}{a vector of probabilities with values in \eqn{[0,1]} to generate the corresponding quantiles.}

\item{predict_sam}{logical; if true, posterior predictive samples for each new location are returned.}

\item{nonref_par_labels}{a vector of numeric values of length \eqn{m}, indicating which partition
each new location belongs to.}

\item{compute_control}{This argument will be used in the future to specific parameters for
generating samples from a copula (e.g., Gumbel) when numerical method is needed.
Currently, the desired accuracy (convergence tolerance) is 1e-5, and
the maximum number of iterations is 1000.}

\item{verbose}{logical; if true, progress of the prediction is printed to the screen.}

\item{nreport}{If \code{verbose = TRUE}, \code{n_report} defines the interval to report
the progress of prediction.}

\item{...}{additional arguments. No additional arguments are supported currently.}
}
\value{
The return object is a list that comprises:
}
\description{
This function produces posterior predictive samples for a set of new locations using an NNMP.
}
\references{
Zheng, X., Kottas, A., and Sansó, B. 
"Nearest-neighbor mixture models for non-Gaussian spatial processes". 
Bayesian Anal. 18 (4) 1191 - 1222, December 2023.
DOI: \href{https://doi.org/10.1214/23-BA1405}{https://doi.org/10.1214/23-BA1405}.
}
\author{
Xiaotian Zheng \email{xiaotianzheng42@gmail.com}
}
