% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarize}
\alias{summarize}
\title{Obtain posterior estimates}
\usage{
summarize(x, pt, ci = c(0.025, 0.975), k = 2)
}
\arguments{
\item{x}{a vector or matrix of posterior samples. If \code{x} is a matrix, 
each row of the matrix corresponds to posterior samples of an unknown parameter.}

\item{pt}{a quoted keyword that specifies point estimate.
Supported keywords are \code{"mean"} and \code{"median"}.}

\item{ci}{a numeric vector of lower and upper probabilities with values in \eqn{[0,1]} 
to generate a credible interval for the unknown parameter(s).}

\item{k}{an integer indicating the number of decimal places to be used.}
}
\value{
A vector of characters. 
        Each element of the vector is in the form: point estimate (lower percentile, upper percentile).
}
\description{
This function calculates point and interval estimates of unknown parameters
using their posterior samples
}
